pyproject.toml
[tool.poetry]
name = "ai-diffusers"
version = "0.1.0"
description = ""
authors = ["Ryo Kuroyanagi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
diffusers = {extras = ["torch"], version = "^0.27.0"}
torch = {version = "^2.0.1+cu117", source = "torch_cu117"}
transformers = "^4.38.2"
anthropic = "^0.20.0"
python-dotenv = "^1.0.1"


[[tool.poetry.source]]
name = "torch_cu117"
url = "https://download.pytorch.org/whl/cu117"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
